Using the scripted to build the variation
Started off making the menu, but then I was given the advice to focus on the game. A wise advice from Prof.Pippin Baar start simple. Started with the hyperspeed, but using the script to make and see the result. Also changed the mouse to keyboard in order to have the game move faster