Added repel function
Added a function in which when the tongue gets closer to the fly, the fly moves higher and run aways. I also changed the function of the flytongue overlap in order to have the tongue continue to go higher when the tongue eats the flies. It allows for consistency and always the player to have a way to eat the fly. Without hating the creator :D