Added the dialogue time
I want the dialogue to disappear after 2 second or so. Also change the movetongue function so that it works with the dialogue. But also kept the old one. But works. In the fly overlap added tongue is inbound,  but commented to see which is more smooth. The tongue getting inbound after touch a fly or if the tongue continues it travels as it eats and makes it tongue way to the top