Added the variable for hyperspeed
The game will be change. New idea is the game is fast but there will be 3 important factors. One the golden will be beyond fast. If it is caught you win. However the silverfly is the poison in the game. If caught the game will initiated ridiculous hyperspeed. Dialogue is needed to inform player hyperspeed is initiated and the mode wiill change