Updated the functions and verified that everyting works
Everything works and all the state works as well as the reset. When normal fly is caught score increase and frog glows. Silver initiates hypermode and in hypermode if you touch silver it cause you to lose. Golden fly is  an automatic win
All other effect sparkles and goldentrails work. Learned that color can be use for some function and not fill