Change the size of the frog and added keys
Forgot to commit when frog sizes was change. I want them to be smaller as they will be assisting the main frog. I added the key function for the border frog I want to make a way for more people to join or if people want to just have fun with help by themselves