Done Patience is key
Added more flies to make winning a bit more realistic. Added the audio to be played when button is pressed. It's more consistent. Reduce the tongue speed, I forgot to change it back. plus it's called patience is key. Move the create fly outside of the setup to allow the restart to restart everything. it's the best way to have the flies respawn. Removed unnecessary comments